[monodroid] More C++ warnings cleanup & C++17 adoption (#3574)
A small cleanup which deals with a handful of remaining warnings
after enabling more extensive diagnostics in our compilers.

Deal with **read**(2) definition differences between Android+Unix
compilers and MinGW: on *nix the third argument, `count`, is defined
to be of type `size_t` while on MinGW it's an `unsigned int`, which
caused an integer conversion warning when building for Windows.

Take advantage of the C++17 `if constexpr ()` construct (works
similar to the preprocessor `#if{n}def` but is part of the language
standard and is type-safe) as well as the more compact (and less of
an eyesore) namespace declaration; instead of:

	namespace xamarin { namespace android { namespace internal

we can now use

	namespace xamarin::android::internal

C++17 allows one to declare and define certain static `constexpr`
members directly in the header file, without the need to add the
definition of such member in a source file somewhere.  Take advantage
of that.