[tests] Check apk tests sizes to avoid regressions (#3594)
Fixes: https://github.com/xamarin/xamarin-android/issues/3593

Added new `<ProcessApkSizes/>` task based on the `<ProcessPlotInput/>`
task, which takes additional argument with apk sizes reference file.
The file is used to check the current apk sizes and fails when the
size difference is over a given threshold; we use 100KiB now. There's
one additional threshold to warn about smalled differences, with
current threshold 50KiB.

The reference files can be update with

	> make update-apk-sizes-reference CONFIGURATION=<Debug or Release>

which copies the current apk sizes files to `tests/apk-sizes-reference`
directory, these are then used by `<ProcessApkSizes/>` task.

The apk sizes files were split by configuration and contain
`$(Configuration)` in the name.  The reason for that is an easier
update of the references after failed PR build, after the size
difference is checked to be valid and desired.  This way it is
possible to just download the results artifacts file and copy the
sizes recorded by CI build into the reference directory.