Address review comments
Remove the `$(_HasCommercialFiles)` conditions from the Android API
documentation items because the files are now part of the open source
build.

I locally verified that the `Mono.Android.xml` file is now included in
the `.vsix` installer and that the `MonoAndroid-*` files are now
included in the `.pkg` installer for an open source build.  I also
compared the file lists from those installers against the previous PR
build and verified that there weren't any other unintended changes to
the installer contents.