[monodroid] Allow loading assemblies from memory (#3365)
Allow execution within the [Xamarin.Android Designer][0] to load
assemblies *from memory* instead of from the filesystem.  This allows
more flexibility in how assemblies are acquired and avoids a
complicated dance of file path juggling and file copying that are
detrimental to performance on Windows and contributes to file-locking
problems.

This is achieved by introducing a new
`xamarin::android::internal::InMemoryAssemblies` type which, like
`xamarin::android::internal::EmbeddedAssemblies`, can be queried when
a new `.dll` needs to be loaded to fetch it from a store that's keyed
per-assembly name and per-AppDomain.  If not found in this new store
it falls back to the existing logic that can load assemblies directly
from disk.

Additionally, fix `mono.android.Runtime.register()` invocations to
take the current AppDomain into consideration.  With the addition of
in-memory assemblies, the following error started appearing:

	Unhandled Exception:
	 System.ArgumentException: Handle value is not valid.
	 Parameter name: method
	  at Java.Interop.JniEnvironment+InstanceMethods.CallObjectMethod (Java.Interop.JniObjectReference instance, Java.Interop.JniMethodInfo method)
	  at Android.Runtime.JNIEnv.CallObjectMethod (System.IntPtr jobject, System.IntPtr jmethod)
	  at Java.Interop.TypeManager.GetClassName (System.IntPtr class_ptr)
	  at Android.Runtime.JNIEnv.RegisterJniNatives (System.IntPtr typeName_ptr, System.Int32 typeName_len, System.IntPtr jniClass, System.IntPtr methods_ptr, System.Int32 methods_len)

This strange error actually stems from the fact that
`TypeManager.GetClassName()` is running in an AppDomain context where
`JNIEnv.Initialize()` has not been called, and thus the static field
`JNIEnv.mid_Class_getName` is `IntPtr.Zero`.

This was puzzling because we do initialize the runtime correctly when
creating a new AppDomain, however after investigating further the
reason turned out to be that `JNIEnv.RegisterJniNatives()` was indeed
being called via a method handle that came from a `JNIEnv`'s
`MonoClass` that was not in the correct AppDomain, thus causing the
mismatch.

The previous code would rely on the fact that on desktop the
`registerType` static global variable would be overwritten every time
the Android runtime was initialized via `Runtime.createNewContext()`,
`Runtime.init()`, and `Runtime.initInternal()`.  This is wrong because
the `MonoMethod` representing `JNIEnv.RegisterJniNatives()` needs to
be changed to point at the version that's loaded in the current
AppDomain, just as we reset the `TypeManager` every time we change the
active context with e.g. `Runtime.switchToContext()`.

The reason that it broke now is because loading the assembly from
disk was actually working around the problem previously since the
assembly would end up being shared across domains, thus the
`registerType` method handle would work fine no matter which domain
it was being executed in.  Switching to in-memory loading and thus
individual assemblies revealed the problem.

[0]: https://docs.microsoft.com/en-us/xamarin/android/user-interface/android-designer