Enable the new targets on Windows
Remove the `$(HostOS)` `Condition` from `AssembleApiDocs` to enable
building the API docs on Windows.

Update the `<Exec/>` task invocations to work on both macOS and Windows
by using `$(ManagedRuntime)` to invoke the version of `mdoc.exe` that's
in the build tree.

Include an extra `\` at the end of `%(SourceDir)` in the `<Exec/>`
tasks.  Without this, `mdoc.exe` would hit an `Illegal characters in
path` exception in `System.IO.Path.LegacyNormalizePath()` on Windows
because [`CommandLineToArgvW()`][0] treats `\` as an escape character
when it precedes `"`.  Adding an additional `\` is a little easier than
using property functions to remove the original `\`.

`mdoc.exe` depends on `ICSharpCode.SharpZipLib.dll` for the `assemble`
subcommand, so add `ICSharpCode.SharpZipLib.dll` to the
`MonoUtilityFile` list in `xaprepare`.  This means that
`ICSharpCode.SharpZipLib.dll` will be added to the bundle during `make
prepare` on macOS and extracted alongside `mdoc.exe` when an existing
bundle is used on macOS or Windows.

Now that `ICSharpCode.SharpZipLib.dll` is part of the bundle, it is
available for inclusion in the open source installer packages, so move
the corresponding `<_MSBuildFiles>` item out from the
`$(_HasCommercialFiles)` condition.

Add a trailing slash on the value of `$(MSBuildSrcDir)` so that the new
targets can follow the convention that the value of a directory path
property always includes a trailing slash.  One reason for this
convention is that if the value of a path item does not have a trailing
slash, then the last part of the path is assigned to the `%(Filename)`
metadata rather than being part of the `%(Directory)` metadata.

Rename `SourceFolder` to `SourceDir` because "Dir" is more common than
"Folder" both in this file and in the built-in MSBuild tasks like
`<MakeDir/>` and `<RemoveDir/>`.

[0]: https://docs.microsoft.com/windows/win32/api/shellapi/nf-shellapi-commandlinetoargvw