[build] Include classlib docs in OSS installers (#3601)
Fixes: https://github.com/xamarin/xamarin-android/issues/3086

Bumps to xamarin/android-api-docs/master@25aa18c2

The [Xamarin.Android API docs were open-sourced][0]!

This means that the public `xamarin-android` repo can now easily
reference the API docs, which in turn means that the OSS "flavors" of
the [macOS `.pkg` and Windows `.vsix` installers][1] can likewise
include and redistribute documentation.

Migrate the `make assemble-docs` build target and its dependencies
from the `docs.make` file in monodroid to a set of MSBuild targets in
xamarin-android.

Add a trailing slash on the value of `$(FrameworkSrcDir)` so that the
new targets can follow the convention that the value of a directory
path property always includes a trailing slash.

Add new MSBuild targets to replace the old `docs.make` targets.
These targets run [`mdoc`][2] to generate ["msxdoc"][3] Microsoft XML
Documentation format files that are used in the `.vsix` installer for
Windows as well, as an [`mdoc assemble`][4]-formatted
`MonoAndroid-lib.zip` archive and `.tree` sfile that are used in the
`.pkg` installer for macOS.

Bring in xamarin/android-api-docs repo as a submodule, set at the
same commit as the corresponding submodule in monodroid.

There are a few intentional differences compared to the original
`docs.make` targets:

  * The Microsoft XML Documentation files are output to the directory
    of the latest *stable* framework version rather than the latest
    *known* framework version.  This is useful because it means the
    MSBuild targets don't need access to the `$(FRAMEWORKS)` value.

  * The OpenTK-1.0 docs are only generated for the `mdoc assemble`
    step and not for the `mdoc export-msxdoc` step.
    `mdoc export-msxdoc` output isn't needed at the moment because
    the `.vsix` installer instead uses the `.xml` file that's
    generated when `Mono.Android.csproj` builds `OpenTK-1.0.csproj`
    for `$(TargetFrameworkVersion)`=`v4.4`.

  * The OpenTK docs are omitted as OpenTK has been deprecated and
    removed from Xamarin.Android.

  * The `make pimpup-docs` target is omitted.  I checked the build
    logs for a recent commercial build and saw that while the build
    does compile `doc-relcontent-merger.exe` and
    `recipe-relcontent-generator.exe`, it never runs either one.
    Starting in [monodroid@c83b93e2][5], the `$(PIMP_STAMP)` target
    that used to run `doc-relcontent-merger.exe` was no longer needed
    and was disabled.

  * The `docs/netdocs.zip` and `docs/mono-libs.zip` targets are also
    intentionally omitted.  Although commercial builds do still run
    those targets, the outputs are not used for anything.

    Based on the contents of the `push-docs` target in `docs.make` and
    the fact that that target was last edited in 2012, I believe the
    `netdocs.zip` and `mono-libs.zip` archives were for an old online
    documentation system that was retired long ago.

I verified that the `Mono.Android.xml` Microsoft XML Documentation
file generated by these new targets was identical to the version
generated by the `docs.make` targets.  Similarly, I verified that the
new monodoc `MonoAndroid-lib.zip` archive generated by these targets
contained the same list of `xml.summary.*` files as the `.zip`
archive generated by the `docs.make` targets.

Additionally, include `ICSharpCode.SharpZipLib.dll` into the mono
bundle (fbfd676c, 29568117) and the macOS `.pkg` and Windows `.vsix`
installers, as `mdoc assemble` requires it.

Add a trailing slash on the value of `$(MSBuildSrcDir)` so that the
new targets can follow the convention that the value of a directory
path property always includes a trailing slash.  One reason for this
convention is that if the value of a path item does not have a
trailing slash, then the last part of the path is assigned to the
`%(Filename)` metadata rather than being part of the `%(Directory)`
metadata.

Rename `%(SourceFolder)` to `%(SourceDir)`, because "Dir" is more
common than "Folder" both in `create-installers.targets` and in the
built-in MSBuild tasks like `<MakeDir/>` and `<RemoveDir/>`.

Finally, include an extra `\` at the end of `%(SourceDir)` in the
`<Exec/>` tasks.  Without this, `mdoc.exe` would hit an
`Illegal characters in path` exception in
`System.IO.Path.LegacyNormalizePath()` on Windows, because
[`CommandLineToArgvW()`][6] treats `\` as an escape character when it
precedes `"`.  Adding an additional `\` is a little easier than using
property functions to remove the original `\`.

[0]: https://devblogs.microsoft.com/xamarin/xamarin-api-docs-open-source-available-now/
[1]: https://github.com/xamarin/xamarin-android/blob/02f3258ade5c12cecfc64cedcf479d2b3bfc41e1/Documentation/workflow/JenkinsBuildArtifacts.md
[2]: http://docs.go-mono.com/?link=man%3amdoc%281%29
[3]: https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/compiler-options/doc-compiler-option
[4]: http://docs.go-mono.com/?link=man%3amdoc-assemble%281%29
[5]: https://github.com/xamarin/monodroid/commit/c83b93e20b0e6f80af7f8e19a6f1dc73f187e4f0
[6]: https://docs.microsoft.com/windows/win32/api/shellapi/nf-shellapi-commandlinetoargvw