[Xamarin.Android.Build.Tasks] deprecate <GetAdditionalResourcesFromAssemblies/>
The `<GetAdditionalResourcesFromAssemblies/>` MSBuild task is the
precursor to Xamarin.Build.Download. It is only used by very old
versions of the Xamarin support library components.

We want to deprecate its behavior in an effort to remove complexity
and improve build performance. We have to scan each assembly for these
assembly-level attributes.

I added a new `XA0121` warning if a library is encountered during a
build that uses these attributes:

* `IncludeAndroidResourcesFromAttribute`
* `NativeLibraryReferenceAttribute`
* `JavaLibraryReferenceAttribute`

I added `[Obsolete]` on these as well.