[tests/MSbuildDeviceIntegration] Improve TimeZone Tests
Currently the TimeZone unit tests do NOT restore
the device TimeZone back to its initial state.
This could effect other tests if they are expected
to run in a particular TimeZone. So we should restore
the state as much as possible.

Also the TimeZone tests REQUIRE the use of

	adb shell su root xxxx

in order to set the TimeZone. This will not work on
any device which has not been rooted. Fortunately we
can check the `ro.debuggable` to see if the device
can run the above command. If it cannot then we should
ignore the test with a suitable message.

Also fixes an issue where if you use the `ADB_TARGET` environment
variable it was ignored when running `RunAdbCommand`.