New ZIP routines for embedded assembly handling
Context: https://github.com/xamarin/xamarin-android/pull/3606

Xamarin.Android stores uncompressed assemblies inside the APK archive in order
to be able to map them to the device's memory and hand the pointer to Mono,
without having to unpack the files from the APK. We've been using an embedded
copy of the `unzip` utility in order to access the archive. Not only the code
was old, it also did much more than was necessary for our purposes.

PR #3606 implemented both the ZIP handling portion as well as had changes to
delay mmapping of the files embedded in APK till the moment they were actually
needed by the Mono runtime. However, some of the tests using the debugger would
either hang or fail to work properly with #3606 and so we decided to split it up
into two parts - zip handling and mmap code. This commit implements the former.

ZIP handling code added here makes a number of assumptions about the state of
the archive:

  * APK is a valid ZIP archive
    Rationale: Android verifies APK/AAB archives and if they weren't valid our
    code wouldn't be executed at all.
  * APK is not an empty ZIP archive
    Rationale: if it were empty, we would have a rather hard time running our
	code from it
  * APK does not use ZIP-64 extensions
    Rationale: for the moment Google use ZIP archives without the 64-bit
	extensions and there's no sign of that changing. If ever the change is made,
	we can easily adjust our code.

Additionally, this commit adds a simple C++ guard class useful in managing
dynamically allocated memory within a function, ensuring that it is
automatically freed on function exit.

Performance impact with our Xamarin.Forms integration test, Android 10 on Pixel
3 XL. Measurements are in milliseconds. Despite omitting the lazy mmaping code,
this commit is nearly as fast as #3606, however the cost of mmaping is not zero
and it will scale up with the number of assemblies in the APK.

Figures represent Xamarin.Android total init time only:

|  **Assembly preload** |  **master** | **this commit** | **Architecture** |
|-----------------------|------------ |-----------------|------------------|
| **enabled**           | 271.373     | 215.785         | armeabi-v7a      |
| **disabled**          | 238.404     | 182.055         | armeabi-v7a      |
| **enabled**           | 255.984     | 197.586         | arm64-v8a        |
| **disabled**          | 225.933     | 169.131         | arm64-v8a        |