[tests] DeviceTest captures & records screenshots
Context: https://dev.azure.com/DevDiv/DevDiv/_build/results?buildId=3075928&view=ms.vss-test-web.test-result-details

I was originally using this in #3649, but it seemed better to do this
in a separate PR.

If a `DeviceTest` fails, it can take a screenshot, add an attachment
to the failed test, and archive it as an artifact. This will help
troubleshoot if things like a button press fails, and if we add more
tests like this down the road.

I used `/data/local/tmp/screenshot.png` as this path has been used by
`Fast Deployment` for a long time and Android hasn't changed it.