[monodroid] New ZIP routines for embedded assembly handling (#3660)
Context: https://github.com/xamarin/xamarin-android/pull/3606

Xamarin.Android stores uncompressed assemblies inside the `.apk`
archive in order to be able to map them to the device's memory and
hand the pointer to Mono, without having to unpack the files from
the `.apk`; see c1956835.  We've been using an embedded copy of the
`unzip` utility in order to read the `.apk`.  Not only was the code
old, it also did much more than was necessary for our purposes,
incurring unnecessary process startup overhead.

PR #3606 implemented both the ZIP handling portion as well as had
changes to delay `mmap()`ing of the files embedded within the `.apk`
until the moment they were actually needed by the Mono runtime.
However, some of the tests using the debugger would either hang or
fail to work properly with #3606 and so we decided to split it up
into two parts: zip handling and mmap code.

This commit implements the former.

Zip handling code added here makes a number of assumptions about the
state of the archive:

  * `.apk` is a valid Zip archive
    Rationale: Android verifies `.apk`/`.aab` archives and if they
    weren't valid our code wouldn't be executed at all.
  * `.apk` is not an empty Zip archive
    Rationale: if it were empty, we would have a rather hard time
    running our code from it
  * `.apk` does not use [ZIP64 extensions][0]
    Rationale: for the moment Google use ZIP archives without the
    64-bit extensions and there's no sign of that changing.  If ever
    the change is made, we can easily adjust our code.

Additionally, this commit adds a simple C++ guard class useful in
managing dynamically allocated memory within a function, ensuring
that it is automatically freed on function exit.

The new Zip handling code reduces the `Runtime.init: end, total time`
measurements.

Times for the `tests/Xamarin.Forms-Performance-Integration` app built
in Release configuration, on a Pixel 3 running Android 10:

| **Assembly preload**  | **master**  | **this commit** | **Architecture** |
|-----------------------|-------------|-----------------|------------------|
| **enabled**           | 271.373     | 215.785         | armeabi-v7a      |
| **disabled**          | 238.404     | 182.055         | armeabi-v7a      |
| **enabled**           | 255.984     | 197.586         | arm64-v8a        |
| **disabled**          | 225.933     | 169.131         | arm64-v8a        |

Despite omitting the lazy `mmap()`ing code, this commit is nearly as
fast as #3606, however the cost of `mmap()` is not zero and it will
scale up with the number of assemblies in the `.apk`.

[0]: https://en.wikipedia.org/wiki/Zip_(file_format)#ZIP64