[tests] DeviceTest captures & records screenshots (#3673)
If a `DeviceTest` fails, it can take a screenshot, add an attachment
to the failed test, and archive it as an artifact.  This will help
troubleshoot if things like a button press fails, and if we add more
tests like this down the road.

I used `/data/local/tmp/screenshot.png` as this path has been used by
`Fast Deployment` for a long time and Android hasn't changed it.