[MSBuildDeviceIntegration] Harden checks for the debugger.
We hit a problem where the text for the log message
for the debugger was changed. It went from

	monodroid-debug: Trying to initialize the debugger with options:

to

	monodroid-debug: mono_runtime_init: Trying to initialize the debugger with options:

Which stopped our unit test matching the line.
So lets make this use just the text

	Trying to initialize the debugger with options:

and make sure we comment the line to if it changes people
also go an update the unit test.