[ci] Add Windows MSBuild test job (#3643)
We'd like to run most if not all of our tests against a complete
`.vsix` or `.pkg` installer in order to better replicate our users
development environments.  All of our 'Test' jobs on Azure Pipelines
are currently configured to test against our installers, however
these jobs are only running on macOS environments.

We'll start to extend our test coverage on Windows by copying the
MSBuild test steps which are in the 'Win Build and Test' stage into
a new **Test** > **MSBuild - Windows** job.

Additionally, perform a bit of refactoring to
`Builder.FrameworkLibDirectory` and `Builder.AndroidMSBuildDirectory`
to support testing against local build trees and system installations
on both macOS and Windows.