[Xamarin.Android.Build.Tasks] Allow `file:` for bundletool ks-pass
Fixes #3700

Commit 04a43082 added code to allow users to use `file:` and `env:`
for the keystore passwords. However we didn't add this support for
the `aab` bundle. This commit adds that support.

It is worth noting that `bundletool` does not support `env:` only
`pass:` and `file:`. So we also need to update the documentation to
mention that.

We also need to update our unit tests to make sure we test
`apksigner`, `jarsigner` and `bundletool`. However in order to
test `bundletool` signing we need to have a device. So we need
to move the `TestAndroidStoreKey` unit test over to the `InstallTests`
class.

There was also a bug in our `StringAssertEx.Contains`. It used
`Assert.Pass`. It turns out that `Assert.Pass` throws an exception
when it is called. As a result NONE of the code after an `Assert.Pass`
will run. So we have a number of tests which were exiting early.
This commit fixes that so that we use `Assert.IsTrue` instread.